pyproject.toml
[project]
name = "todoist-ai"
version = "0.1.0"
description = "Pull my todoist tasks and feed them into an AI, maybe it will help?"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "py-jsonl>=1.3.10",
    "pydantic>=2.10.6",
    "todoist-api-python>=2.1.7",
]
